% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\docType{data}
\name{bmi}
\alias{bmi}
\title{FUSION output on 2015 BMI GWAS}
\format{See FUSION \url{http://gusevlab.org/projects/fusion/}}
\usage{
bmi
}
\description{
FUSION output on 2015 BMI GWAS
}
\keyword{datasets}
