% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_cor}
\alias{get_cor}
\title{Compute \eqn{\rho_ge}}
\usage{
get_cor(t1, t2, regions = grch37.eur.loci, min_regions = 10,
  nsamples = 100)
}
\arguments{
\item{t1}{data.frame-like, TWAS results for trait1}

\item{t2}{data.frame-like, TWAS results for trait2}

\item{regions}{data.frame-like, Approximate independent regions. Requires columns (Chr, Start, Stop). Default is estimated blocks in Europeans.}

\item{min_regions}{integer, Minimum number of independent regions for inference}

\item{nsamples}{integer, Number of samples to perform for bootstrapping}
}
\value{
data.frame with estimates
}
\description{
Compute \eqn{\rho_ge}
}
