% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_independent}
\alias{get_independent}
\title{Get spatially independent genes}
\usage{
get_independent(genes, regions)
}
\arguments{
\item{genes}{data.frame-like, gene information for a TWAS}

\item{regions}{data.frame-like, independent regions contained in Chr, Start, Stop variables}
}
\value{
ind_genes data.frame-like, data.frame annotated with independent blocks for each gene
}
\description{
Get spatially independent genes
}
