% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_trait}
\alias{get_trait}
\title{Get FUSION table}
\usage{
get_trait(twas_tbl, ngwas)
}
\arguments{
\item{twas_tbl}{data.frame-like, Table/Dataframe of TWAS results from FUSION}

\item{ngwas}{integer, Sample size for GWAS}
}
\value{
Annotated data.frame suitable for RHOGE inference
}
\description{
Get FUSION table
}
