% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch37.asn.loci.R
\docType{data}
\name{grch37.asn.loci}
\alias{grch37.asn.loci}
\title{Apprxoimately independent linkage-disequilibrium blocks for Asian ancestry samples under GRCh37 coordinaSTOP}
\format{A data frame with 1445 rows and 3 variables:
\describe{
  \item{CHR}{Chromosome}
  \item{START}{Start of LD block}
  \item{STOP}{End of LD block}
}}
\source{
\url{https://academic.oup.com/bioinformatics/article/32/2/283/1743626}
}
\usage{
grch37.asn.loci
}
\description{
Apprxoimately independent linkage-disequilibrium blocks for Asian ancestry samples under GRCh37 coordinaSTOP
}
\keyword{datasets}
