% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{rhoge.gw}
\alias{rhoge.gw}
\title{Calculates genome-wide genetic correlation between two complex traits using TWAS summary statistics.}
\usage{
rhoge.gw(trait1, trait2, n1, n2, p = 0.05, regions = grch37.eur.loci)
}
\arguments{
\item{trait1}{data.frame-like, TWAS results for trait 1}

\item{trait2}{data.frame-like, TWAS results for trait 2}

\item{n1}{integer, sample-size for trait 1 GWAS}

\item{n2}{integer, sample-size for trait 2 GWAS}

\item{p}{double, nominal significance threshold for genome-wide RhoGE estimate (default=0.05)}

\item{regions}{data.frame-like, Approximate independent regions. Requires columns (Chr, Start, Stop). Default is estimated blocks in Europeans.}
}
\description{
Calculates genome-wide genetic correlation between two complex traits using TWAS summary statistics.
}
