% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg.R
\docType{data}
\name{tg}
\alias{tg}
\title{FUSION output on 2013 Triglycerides GWAS}
\format{See FUSION \url{http://gusevlab.org/projects/fusion/}}
\usage{
tg
}
\description{
FUSION output on 2013 Triglycerides GWAS
}
\keyword{datasets}
