% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{validate_pval}
\alias{validate_pval}
\title{Internal pval validation}
\usage{
validate_pval(p, name)
}
\arguments{
\item{p}{double, nominal significance threshold for genome-wide RhoGE estimate (default=0.05)}

\item{name}{name of variable}
}
\description{
Internal pval validation
}
