% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{validate_size}
\alias{validate_size}
\title{Internal parameter validation}
\usage{
validate_size(n, name)
}
\arguments{
\item{n}{size}

\item{name}{name of variable}
}
\description{
Internal parameter validation
}
