% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{validate_tables}
\alias{validate_tables}
\title{Internal data.frame/table validation}
\usage{
validate_tables(trait1, trait2, regions = grch37.eur.loci)
}
\arguments{
\item{trait1}{data.frame-like, containing TWAS results for trait 1}

\item{trait2}{data.frame-like, containing TWAS results for trait 2}

\item{regions, }{data.frame-like containing approximate independent regions. Requires columns (Chr, Start,}
}
\description{
Internal data.frame/table validation
}
